Makes NUXButton configurable using UIAppearance
Removes NUXSubmitButton class. Removes button background assets and generates button background images.
