--- WordPressAuthenticator/NUX/NUXSubmitButton.swift
@@ -1,293 +1,0 @@
-import UIKit
-import WordPressShared
-
-let NUXSubmitButtonDisabledAlpha = CGFloat(0.25)
-
-/// A protocol for an element that can display a UIActivityIndicatorView
-@objc
-public protocol ActivityIndicatorButton {
-    func showActivityIndicator(_ show: Bool)
-}
-
-/// A stylized button used by NUX controllers. The button presents white text
-/// surrounded by a white border.  It also can display a `UIActivityIndicatorView`.
-///
-@objc
-open class NUXSubmitButton: UIButton, ActivityIndicatorButton {
-
-    @objc var isAnimating: Bool {
-        get {
-            return activityIndicator.isAnimating
-        }
-    }
-
-    @objc let activityIndicator: UIActivityIndicatorView = {
-        let indicator = UIActivityIndicatorView(activityIndicatorStyle: .white)
-        indicator.hidesWhenStopped = true
-        return indicator
-    }()
-
-    override open func layoutSubviews() {
-        super.layoutSubviews()
-
-        if activityIndicator.isAnimating {
-            titleLabel?.frame = CGRect.zero
-
-            var frm = activityIndicator.frame
-            frm.origin.x = (frame.width - frm.width) / 2.0
-            frm.origin.y = (frame.height - frm.height) / 2.0
-            activityIndicator.frame = frm
-        }
-    }
-
-    // MARK: - Instance Methods
-
-
-    /// Toggles the visibility of the activity indicator.  When visible the button
-    /// title is hidden.
-    ///
-    /// - Parameter show: True to show the spinner. False hides it.
-    ///
-    open func showActivityIndicator(_ show: Bool) {
-        if show {
-            activityIndicator.startAnimating()
-        } else {
-            activityIndicator.stopAnimating()
-        }
-        configureBackgrounds()
-        configureTitleColors()
-        setNeedsLayout()
-    }
-
-    func didChangePreferredContentSize() {
-        titleLabel?.adjustsFontForContentSizeCategory = true
-    }
-
-
-    // MARK: UIAppearance Customizations
-    /// Style: Primary + Normal State
-    ///
-    @objc public dynamic var primaryNormalBackgroundColor = Primary.normalBackgroundColor {
-        didSet {
-            configureBackgrounds()
-        }
-    }
-    @objc public dynamic var primaryNormalBorderColor = Primary.normalBorderColor {
-        didSet {
-            configureBackgrounds()
-        }
-    }
-
-    /// Style: Primary + Highlighted State
-    ///
-    @objc public dynamic var primaryHighlightBackgroundColor = Primary.highlightBackgroundColor {
-        didSet {
-            configureBackgrounds()
-        }
-    }
-    @objc public dynamic var primaryHighlightBorderColor = Primary.highlightBorderColor {
-        didSet {
-            configureBackgrounds()
-        }
-    }
-
-    /// Style: Secondary
-    ///
-    @objc public dynamic var secondaryNormalBackgroundColor = Secondary.normalBackgroundColor {
-        didSet {
-            configureBackgrounds()
-        }
-    }
-    @objc public dynamic var secondaryNormalBorderColor = Secondary.normalBorderColor {
-        didSet {
-            configureBackgrounds()
-        }
-    }
-    @objc public dynamic var secondaryHighlightBackgroundColor = Secondary.highlightBackgroundColor {
-        didSet {
-            configureBackgrounds()
-        }
-    }
-    @objc public dynamic var secondaryHighlightBorderColor = Secondary.highlightBorderColor {
-        didSet {
-            configureBackgrounds()
-        }
-    }
-
-    /// Style: Disabled State
-    ///
-    @objc public dynamic var disabledBackgroundColor = Disabled.backgroundColor {
-        didSet {
-            configureBackgrounds()
-        }
-    }
-    @objc public dynamic var disabledBorderColor = Disabled.borderColor {
-        didSet {
-            configureBackgrounds()
-        }
-    }
-
-    /// Style: Title!
-    ///
-    @objc public dynamic var titleFont = Title.defaultFont {
-        didSet {
-            configureTitleLabel()
-        }
-    }
-    @objc public dynamic var primaryTitleColor = Title.primaryColor {
-        didSet {
-            configureTitleColors()
-        }
-    }
-    @objc public dynamic var secondaryTitleColor = Title.secondaryColor {
-        didSet {
-            configureTitleColors()
-        }
-    }
-    @objc public dynamic var disabledTitleColor = Title.disabledColor {
-        didSet {
-            configureTitleColors()
-        }
-    }
-
-    /// Insets to be applied over the Contents.
-    ///
-    @objc public dynamic var contentInsets = UIImage.Metrics.contentInsets {
-        didSet {
-            configureInsets()
-        }
-    }
-
-    /// Indicates if the current instance should be rendered with the "Primary" Style.
-    ///
-    @IBInspectable var isPrimary: Bool = false {
-        didSet {
-            configureBackgrounds()
-            configureTitleColors()
-        }
-    }
-
-
-    // MARK: - LifeCycle Methods
-
-    open override func didMoveToWindow() {
-        super.didMoveToWindow()
-        configureAppearance()
-    }
-
-    open override func awakeFromNib() {
-        super.awakeFromNib()
-        configureAppearance()
-    }
-
-
-    /// Setup: Everything = [Insets, Backgrounds, titleColor(s), titleLabel]
-    ///
-    private func configureAppearance() {
-        configureInsets()
-        configureBackgrounds()
-        configureTitleColors()
-        configureTitleLabel()
-    }
-
-    /// Setup: FancyButton's Default Settings
-    ///
-    private func configureInsets() {
-        contentEdgeInsets = contentInsets
-    }
-
-    /// Setup: BackgroundImage
-    ///
-    private func configureBackgrounds() {
-        let normalImage: UIImage
-        let highlightedImage: UIImage
-        let disabledImage = UIImage.renderBackgroundImage(fill: disabledBackgroundColor, border: disabledBorderColor, cornerRadius: Metrics.backgroundCornerRadius)
-
-        if isPrimary {
-            normalImage = UIImage.renderBackgroundImage(fill: primaryNormalBackgroundColor, border: primaryNormalBorderColor, cornerRadius: Metrics.backgroundCornerRadius)
-            highlightedImage = UIImage.renderBackgroundImage(fill: primaryHighlightBackgroundColor, border: primaryHighlightBorderColor, cornerRadius: Metrics.backgroundCornerRadius)
-        } else {
-            normalImage = UIImage.renderBackgroundImage(fill: secondaryNormalBackgroundColor, border: secondaryNormalBorderColor, cornerRadius: Metrics.backgroundCornerRadius)
-            highlightedImage = UIImage.renderBackgroundImage(fill: secondaryHighlightBackgroundColor, border: secondaryHighlightBorderColor, cornerRadius: Metrics.backgroundCornerRadius)
-        }
-
-        setBackgroundImage(normalImage, for: .normal)
-        setBackgroundImage(highlightedImage, for: .highlighted)
-        setBackgroundImage(disabledImage, for: .disabled)
-
-        addSubview(activityIndicator)
-    }
-
-    /// Setup: TitleColor
-    ///
-    private func configureTitleColors() {
-        let titleColorNormal = isPrimary ? primaryTitleColor : secondaryTitleColor
-
-        setTitleColor(titleColorNormal, for: .normal)
-        setTitleColor(titleColorNormal, for: .highlighted)
-        setTitleColor(disabledTitleColor, for: .disabled)
-    }
-
-    /// Setup: TitleLabel
-    ///
-    private func configureTitleLabel() {
-        titleLabel?.font = titleFont
-        titleLabel?.adjustsFontForContentSizeCategory = true
-        titleLabel?.textAlignment = .center
-    }
-}
-
-// MARK: - Nested types
-private extension NUXSubmitButton {
-    /// Style: Primary
-    ///
-    struct Primary {
-        static let normalBackgroundColor = WPStyleGuide.lightBlue()
-        static let normalBorderColor = WPStyleGuide.darkBlue()
-        static let highlightBackgroundColor = WPStyleGuide.darkBlue()
-        static let highlightBorderColor = normalBorderColor
-    }
-
-    /// Style: Secondary
-    ///
-    struct Secondary {
-        static let normalBackgroundColor = UIColor.clear
-        static let normalBorderColor = UIColor.white
-        static let highlightBackgroundColor = UIColor.white
-        static let highlightBorderColor = highlightBackgroundColor
-    }
-
-    /// Style: Disabled
-    ///
-    struct Disabled {
-        static let backgroundColor = UIColor.white
-        static let borderColor = UIColor(white: 1.0, alpha: NUXSubmitButtonDisabledAlpha)
-    }
-
-    /// Style: Title
-    ///
-    struct Title {
-        static let primaryColor = UIColor.white
-        static let secondaryColor = WPStyleGuide.darkBlue()
-        static let disabledColor = UIColor(white: 1.0, alpha: NUXSubmitButtonDisabledAlpha)
-        static let defaultFont = WPFontManager.systemSemiBoldFont(ofSize: 17.0)
-    }
-}
-
-// MARK: -
-//
-extension NUXSubmitButton {
-    override open func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
-        super.traitCollectionDidChange(previousTraitCollection)
-        if previousTraitCollection?.preferredContentSizeCategory != traitCollection.preferredContentSizeCategory {
-            didChangePreferredContentSize()
-        }
-    }
-
-    struct Metrics {
-        static let backgroundCornerRadius = CGFloat(5.0)
-        static let backgroundCapInsets = UIEdgeInsets(top: 5, left: 5, bottom: 5, right: 5)
-        static let backgroundShadowOffset = CGSize(width: 1, height: 1)
-        static var contentInsets = UIEdgeInsets(top: 0, left: 15, bottom: 0, right: 15)
-    }
-}
