--- WordPressAuthenticator/Resources/Assets.xcassets/beveled-secondary-button.imageset/Contents.json
@@ -1,65 +1,0 @@
-{
-  "images" : [
-    {
-      "resizing" : {
-        "mode" : "9-part",
-        "center" : {
-          "mode" : "stretch",
-          "width" : 1,
-          "height" : 1
-        },
-        "cap-insets" : {
-          "bottom" : 10,
-          "top" : 9,
-          "right" : 9,
-          "left" : 9
-        }
-      },
-      "idiom" : "universal",
-      "filename" : "beveled-secondary-button.png",
-      "scale" : "1x"
-    },
-    {
-      "resizing" : {
-        "mode" : "9-part",
-        "center" : {
-          "mode" : "stretch",
-          "width" : 1,
-          "height" : 1
-        },
-        "cap-insets" : {
-          "bottom" : 20,
-          "top" : 18,
-          "right" : 18,
-          "left" : 18
-        }
-      },
-      "idiom" : "universal",
-      "filename" : "beveled-secondary-button@2x.png",
-      "scale" : "2x"
-    },
-    {
-      "resizing" : {
-        "mode" : "9-part",
-        "center" : {
-          "mode" : "stretch",
-          "width" : 1,
-          "height" : 1
-        },
-        "cap-insets" : {
-          "bottom" : 30,
-          "top" : 24,
-          "right" : 24,
-          "left" : 24
-        }
-      },
-      "idiom" : "universal",
-      "filename" : "beveled-secondary-button@3x.png",
-      "scale" : "3x"
-    }
-  ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
